+++ metadata.rb
@@ -4,12 +4,14 @@ maintainer_email  "cookbooks@opscode.com"
 license           "Apache 2.0"
 description       "Installs memcached and provides a define to set up an instance of memcache via runit"
 long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version           "1.1.2"
-depends           "runit"
+version           "1.4.1"
+depends           "runit", "~> 1.0"
+depends           "yum"
 
 recipe "memcached", "Installs and configures memcached"
 
-%w{ ubuntu debian redhat fedora centos scientific amazon }.each do |os|
+%w{ ubuntu debian redhat fedora centos
+    scientific amazon smartos suse }.each do |os|
   supports os
 end