+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## v1.4.0:
+
+### Improvement
+
+- [COOK-2756]: add SUSE support to memcached cookbook
+- [COOK-2791]: Remove the template for Karmic from the memcached cookbook
+
+### Bug
+
+- [COOK-2600]: support memcached on SmartOS
+
+## v1.3.0
+
+* [COOK-2386] - update `memcached_instance` definition for
+  `runit_service` resource
+
+## v1.2.0:
+
+* [COOK-1469] - include yum epel recipe on RHEL 5 (introduces yum
+  cookbook dependency)
+* [COOK-2202] - Fix typo in previous ticket/commits
+* [COOK-2266] - pin runit dependency
+
 ## v1.1.2:
 
 * [COOK-990] - params insite runit_service isn't the same as outside